move Channel to crate::raw_channel::Channel
